Fixed some missed str to int16 conversions for the move from short to int for member ID. Removed hardcoded loop value of 120 for max number of members. Updated error message when adding duplicate members to a term in Attendance.

+++ Marimba/Forms/MemberList.cs
@@ -287,7 +287,7 @@ namespace Marimba.Forms
                 foreach (ListViewItem name in memberlist)
                 {
                     // search the current term for the member
-                    if (ClsStorage.currentClub.listTerms[ClsStorage.currentClub.listTerms.Count - 1].memberSearch(Convert.ToInt16(name.SubItems[5].Text)) >= 0)
+                    if (ClsStorage.currentClub.listTerms[ClsStorage.currentClub.listTerms.Count - 1].memberSearch(Convert.ToInt32(name.SubItems[5].Text)) >= 0)
                         tempList.Add(name);
                 }
                 memberlist = tempList;
